# Phase 7.2: Performance Optimization & Profiling (Issue #88)