… ai-coach-api/src/services/processing_config.rs
… docs/performance-optimization.md